Make the Kotlin convention plugins JVM/Multiplatform-agnostic
Before this change, depending on whether a subproject was JVM-based
or multiplatform, different convention plugins were applied.
Those were written separately and only shared some logic via yet
another file.
Now, all of this is joined into a single file, helping us to
consistently apply whatever changes we make and simplifying the
structure.
The next step is to extract some logic they share into separate
functions, so that the code duplication is lowered.
